# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carvis — Canvas LMS AI assistant
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CanvasPatInput {
  canvasUrl: string;
  pat: string;
}

export interface CanvasUrlInput {
  canvasUrl: string;
}

export interface UserProfile {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  canvasBaseUrl?: string | null;
}

export interface AuthResult {
  success: boolean;
  user?: UserProfile;
}

export interface OAuthStartResult {
  url: string;
}

export interface Me {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  canvasBaseUrl?: string | null;
  canvasConnected?: boolean;
}

export interface Assignment {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  points?: number | null;
  /** @nullable */
  url?: string | null;
  completed: boolean;
  courseId?: string;
  /** @nullable */
  courseName?: string | null;
}

export interface Course {
  id: string;
  name: string;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  lastSynced?: string | null;
  assignments?: Assignment[];
}

export interface UserData {
  user: UserProfile;
  courses: Course[];
  hasData: boolean;
}

export interface SyncResult {
  success: boolean;
  courseCount?: number;
  assignmentCount?: number;
  /** @nullable */
  message?: string | null;
}

export interface GradeItem {
  courseId: string;
  name: string;
  /** @nullable */
  currentScore?: number | null;
  /** @nullable */
  finalScore?: number | null;
  /** @nullable */
  letterGrade?: string | null;
}

export interface GradeList {
  grades: GradeItem[];
}

export interface DashboardSummary {
  upcomingCount: number;
  overdueCount: number;
  completedCount: number;
  totalAssignments: number;
  /** @nullable */
  avgGrade?: number | null;
  nextDue?: Assignment;
  courseCount?: number;
}

export interface VoiceInput {
  text: string;
}

export interface VoiceResult {
  /** @nullable */
  intent?: string | null;
  response: string;
  /** @nullable */
  confidence?: number | null;
}

export interface ConversationEntry {
  id?: number;
  role: string;
  message: string;
  /** @nullable */
  intent?: string | null;
  createdAt: string;
}

export interface ConversationHistory {
  messages: ConversationEntry[];
}

export interface PageElement {
  id: string;
  tag: string;
  text: string;
  /** @nullable */
  ariaLabel?: string | null;
  /** @nullable */
  placeholder?: string | null;
  /** @nullable */
  href?: string | null;
}

export interface PageContext {
  url: string;
  title: string;
  elements: PageElement[];
}

export interface ExtensionAgentInput {
  command: string;
  pageContext: PageContext;
}

export interface AgentAction {
  type: string;
  /** @nullable */
  elementId?: string | null;
  /** @nullable */
  value?: string | null;
  /** @nullable */
  direction?: string | null;
  /** @nullable */
  url?: string | null;
}

export interface ExtensionAgentResult {
  response: string;
  action?: AgentAction;
  /** @nullable */
  blocked?: boolean | null;
}