# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Carvis — Canvas LMS AI assistant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/canvas/start:
    post:
      operationId: authCanvasOauthStart
      tags: [auth]
      summary: Start Canvas OAuth flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasUrlInput"
      responses:
        "200":
          description: OAuth URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthStartResult"
        "400":
          description: Bad request
...
  /voice/command:
    post:
      operationId: voiceCommand
      tags: [voice]
      summary: Process a voice command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceInput"
      responses:
        "200":
          description: Command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceResult"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    CanvasUrlInput:
      type: object
      required: [canvasUrl]
      properties:
        canvasUrl:
          type: string
...
    OAuthStartResult:
      type: object
      required: [url]
      properties:
        url:
          type: string
...
    VoiceInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
    VoiceResult:
      type: object
      required: [response]
      properties:
        intent:
          type: ["string", "null"]
        response:
          type: string
        confidence:
          type: ["number", "null"]
...