# .migration-backup/lib/api-zod/src/generated/types/voiceResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carvis — Canvas LMS AI assistant
 * OpenAPI spec version: 0.1.0
 */

export interface VoiceResult {
  /** @nullable */
  intent?: string | null;
  response: string;
  /** @nullable */
  confidence?: number | null;
}